vision_service/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
opencv-python==4.10.0.84
Pillow==10.4.0
imagehash==4.3.1
numpy==2.1.3
python-multipart==0.0.12
requests==2.32.3
